Minor patch

+++ fetch.py
@@ -158,7 +158,7 @@ def main():
     parser = optparse.OptionParser(version=VERSION)
     parser.add_option("--anonymity", dest="anonymity", help="Regex for filtering anonymity (e.g. \"anonymous|elite\")")
     parser.add_option("--country", dest="country", help="Regex for filtering country (e.g. \"china|brazil\")")
-    parser.add_option("--max-latency", dest="maxLatency", type=int, help="Maximum (tolerable) latency in seconds (default %d)" % TIMEOUT)
+    parser.add_option("--max-latency", dest="maxLatency", type=float, help="Maximum (tolerable) latency in seconds (default %d)" % TIMEOUT)
     parser.add_option("--threads", dest="threads", type=int, help="Number of scanning threads (default %d)" % THREADS)
 
     def _(self, *args):
